requirements.txt
chime==0.7.0
datasets==2.19.0
faiss_cpu==1.8.0
fastapi==0.111.0
gensim==4.3.2
httpx==0.27.0
joblib==1.3.2
matplotlib==3.7.0
nltk==3.6.1
numpy==1.23.5
pandas==1.5.3
pydantic==1.10.12
Requests==2.31.0
scikit_learn==1.4.2
scipy==1.13.0
sentence_transformers==2.7.0
skimpy==0.0.14
spacy==3.7.4
tensorflow_hub==0.16.1
torch==2.3.0
tqdm==4.66.1
uvicorn==0.29.0